+++ libs/ionic-react/src/schematics/application/schema.d.ts
@@ -13,4 +13,5 @@ export interface ApplicationSchematicSchema {
   classComponent?: boolean;
   skipWorkspaceJson?: boolean;
   js?: boolean;
+  disableSanitizer: boolean;
 }